flask_base/requirements.txt
Flask~=3.0.0
Flask-Login
Flask-SQLAlchemy
requests~=2.31.0
werkzeug~=3.0.1

PyYAML
apispec~=6.3.0
marshmallow~=3.20.1
flask-swagger-ui
apispec-webframeworks
SQLAlchemy~=2.0.23
setuptools